requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.34
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
